requirements.txt
# backend dependencies
fastapi
uvicorn[standard]
python-dotenv
google-genai
httpx
supabase
Pillow
python-multipart
pydantic